more audit fixes
